Add hook that signs siww msg and creates session
