+++ dapp/src/hooks/index.ts
@@ -32,3 +32,4 @@ export { default as useLocalStorage } from "./useLocalStorage"
 export { default as useDetectReferral } from "./useDetectReferral"
 export { default as useReferral } from "./useReferral"
 export { default as useMats } from "./useMats"
+export { default as useSignMessageAndCreateSession } from "./useSignMessageAndCreateSession"
